start 1.21.11 porting

+++ src/main/java/org/betterx/betterend/blocks/HydraluxPetalBlock.java
@@ -32,7 +32,7 @@ public class HydraluxPetalBlock extends BaseBlock.Wood {
             @NotNull BlockState blockState,
             @NotNull BlockPos blockPos,
             @NotNull Entity entity,
-            float f
+            double f
     ) {
     }
 }
